part4 completed

+++ part4/blogs/controllers/blogs.js
@@ -1,8 +1,19 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blogs')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
+
+/* const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+} */
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -11,39 +22,55 @@ blogsRouter.get('/:id', async (request, response) => {
   response.json(blog)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.tokenExtractor, middleware.userExtractor, async (request, response) => {
   const body = request.body
+  const user = request.user
 
   const newBlog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes ? body.likes : 0,
+    user: user.id
   })
 
   const blog = await newBlog.save()
+  user.blogs = user.blogs.concat(blog._id)
+  await user.save()
   response.status(201).json(blog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
-  if (!deletedBlog) response.status(404).end()
-  else response.status(204).end()
+blogsRouter.delete('/:id', middleware.tokenExtractor, middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (user.id.toString() !== blog.user.toString()) response.status(401).json({ error: 'unauthorized' })
+  else if (!blog) response.status(404).end()
+  else {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', middleware.tokenExtractor, middleware.userExtractor, async (request, response) => {
   const body = request.body
 
-  const blog = {
+  const newblog = {
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
   }
 
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-  if (!updatedBlog) response.status(404).end()
-  else response.json(updatedBlog)
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (user.id.toString() !== blog.user.toString()) response.status(401).json({ error: 'unauthorized' })
+  else if (!blog) response.status(404).end()
+  else {
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, newblog, { new: true })
+    response.status(200).json(updatedBlog)
+  }
 })
 
 module.exports = blogsRouter
